moves expression functions to own file and collocates types
